% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{confintFitInterval}
\alias{confintFitInterval}
\title{Compute Confidence Intervals for Parameters of a \link{ng}lmer Fit}
\usage{
confintFitInterval(fit)
}
\arguments{
\item{fit}{A fitted \link{ng}lmer model or profile}
}
\value{
Depending on the method specified, confint() computes confidence
intervals by
}
\description{
Compute confidence intervals on the parameters of a *lmer() model fit
(of class"merMod").
}
\author{
juanprimate
}
