% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{contest1DFit}
\alias{contest1DFit}
\title{Test of Contrasts, contest1D}
\usage{
contest1DFit(fit)
}
\arguments{
\item{fit}{An object of class "lmerModLmerTest".}
}
\value{
A data.frame or a list of data.frames. If the design matrix is rank
deficient, lmer drops columns for the aliased coefficients from the design
matrix and excludes the corresponding aliased coefficients from fixef(model).
the estimate to intercept
}
\description{
Tests of vector or matrix contrasts for lmer model fits. Same test, but now
as a t-test instead:
}
\author{
juanprimate
}
