% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{contestFit}
\alias{contestFit}
\title{Test of Contrasts, contest}
\usage{
contestFit(dataset, varY, formula, REML)
}
\arguments{
\item{dataset}{An data frame containing the variables named in formula.}

\item{varY}{An object of class "numeric" (like response).}

\item{formula}{A two-sided linear formula object describing both the
fixed-effects and random-effects part of the model, with the response on
the left of a ~ operator and the terms, separated by + operators, on the
right. Random-effects terms are distinguished by vertical bars (|)
separating expressions for design matrices from grouping factors.}

\item{REML}{A logical scalar - Should the estimates be chosen to optimize
the REML criterion (as opposed to the log-likelihood)?
\enumerate{
\item REML = FALSE is used in case of comparing models with different
“Fixed effects” (during the simplification of model)
\item REML = TRUE is used in case of different random effects on the comparing
models.
}}
}
\value{
A data.frame or a list of data.frames. If the design matrix is rank
deficient, lmer drops columns for the aliased coefficients from the design
matrix and excludes the corresponding aliased coefficients from fixef(model).
the estimate to intercept with confidence interval
}
\description{
Tests of vector or matrix contrasts for lmer model fits. Contrast test
(contest) using a custom contrast:
Here we make the 2-df joint test of the main effects of Gender and
Information
}
\author{
juanprimate
}
