% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{contestMDFit}
\alias{contestMDFit}
\title{Test of Contrasts, contestMD}
\usage{
contestMDFit(fit)
}
\arguments{
\item{fit}{An object of class "lmerModLmerTest".}
}
\value{
A data.frame or a list of data.frames. If the design matrix is rank
deficient, lmer drops columns for the aliased coefficients from the design
matrix and excludes the corresponding aliased coefficients from fixef(model).
the sum square, mean sum square and degree freedom and statistic
}
\description{
Tests of vector or matrix contrasts for lmer model fits. Make the 1-df
F-test of the effect of Days^2:
}
\author{
juanprimate
}
