% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{covarianceMatrix}
\alias{covarianceMatrix}
\title{Computing the covariance matrix}
\usage{
covarianceMatrix(fit)
}
\arguments{
\item{fit}{An object of class "lmerModLmerTest".}
}
\value{
A covariance matrix.
}
\description{
Coercing an lme4::lmer model-object (of class ’lmerMod’) to a model-object
of class ’lmerModLmerTest’
involves computing the covariance matrix of the variance parameters and the
gradient (Jacobian) of
cov(beta) with respect to the variance parameters.
}
\author{
juanprimate
}
