% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semIteracao.R
\name{create_formulaComUma}
\alias{create_formulaComUma}
\title{Create a Formula with Random Effects}
\usage{
create_formulaComUma(input)
}
\arguments{
\item{input}{Una lista que debe contener los siguientes elementos:
\describe{
\item{dataset}{Un data frame con los datos.}
\item{varY}{Nombre de la variable dependiente en el modelo.}
\item{REML}{Un valor lógico que indica si se debe usar REML (TRUE) o máxima verosimilitud (FALSE).}
\item{covariavel}{Nombre de la covariable para el efecto aleatorio.}
}}
}
\value{
Una fórmula de modelo lineal mixto con el efecto aleatorio especificado.
}
\description{
Esta función crea una fórmula para un modelo lineal mixto, especificando un efecto aleatorio.
}
\examples{
input <- list(
  dataset = mi_dataset,
  varY = "mi_variable_dependiente",
  REML = TRUE,
  covariavel = "mi_covariable"
)
formula <- create_formulaComUma(input)
}
\author{
Juanprimate
}
