% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{create_formulaTodo}
\alias{create_formulaTodo}
\title{Create a Random Effects Formula}
\usage{
create_formulaTodo(input)
}
\arguments{
\item{input}{A list with elements:
\itemize{
\item \code{dataset}: A data frame containing the data.
\item \code{varY}: A character string indicating the dependent variable.
\item \code{REML}: A logical value (although it is not used within the function).
}}
}
\value{
A formula object that represents the random effects structure.
}
\description{
This function generates a formula for random effects using the input dataset, dependent variable (\code{varY}),
and any remaining variables in the dataset as random effects. The function combines all possible combinations
of the covariates and constructs a formula that can be used in mixed models.
}
\examples{
dataset <- data.frame(y = rnorm(100), x1 = rnorm(100), x2 = rnorm(100))
input <- list(dataset = dataset, varY = "y", REML = TRUE)
create_formulaTodo(input)

}
\author{
juanprimate
}
