% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semIteracao.R
\name{extract_p_value}
\alias{extract_p_value}
\title{Extract p-value for a Specific Term from ANOVA Results}
\usage{
extract_p_value(df, term)
}
\arguments{
\item{df}{Un marco de datos que contiene resultados de ANOVA, con términos en las filas y p-valores en una columna llamada "Pr(>Chisq)".}

\item{term}{El término para el cual se desea extraer el p-valor.}
}
\value{
El p-valor asociado al término especificado. Si el término no se encuentra en el marco de datos, se devuelve NA.
}
\description{
Esta función extrae el p-valor asociado a un término específico de un marco de datos que contiene resultados de ANOVA.
}
\examples{
# Supongamos que `output$randFitAnovaFit` es el marco de datos con resultados de ANOVA
p_value <- extract_p_value(output$randFitAnovaFit, "employee:batch")
}
\author{
Juanprimate
}
