% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{formulaRandom}
\alias{formulaRandom}
\title{Creating a formula from a string to mixed model White test}
\usage{
formulaRandom(dataset, varY)
}
\arguments{
\item{dataset}{A data set of class "data.frame".}

\item{varY}{An object of class "character".}
}
\value{
An object of class \code{"formula"} which contains a symbolic model
formula.
}
\description{
Creating a formula from a string to random effect
}
\author{
juanprimate
}
