% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{formulaRandomList}
\alias{formulaRandomList}
\title{Creating a formula from a string to mixed model, formulaRandomList}
\usage{
formulaRandomList(data, varY)
}
\arguments{
\item{data}{A data set of class "list".}

\item{varY}{An object of class "character".}
}
\value{
An object of class \code{"formula"} which contains a symbolic model
formula.
}
\description{
Creating a formula from a string to random effect
}
\author{
juanprimate
}
