% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{lmerRandomFit}
\alias{lmerRandomFit}
\title{Fit Linear Mixed-Effects Models}
\usage{
lmerRandomFit(dataset, varY, formula, REML)
}
\arguments{
\item{dataset}{An data frame containing the variables named in formula.}

\item{varY}{An object of class "numeric" (like response).}

\item{formula}{A two-sided linear formula object describing both the
fixed-effects and random-effects part of the model, with the response on
the left of a ~ operator and the terms, separated by + operators, on the
right. Random-effects terms are distinguished by vertical bars (|)
separating expressions for design matrices from grouping factors.}

\item{REML}{A logical scalar - Should the estimates be chosen to optimize
the REML criterion (as opposed to the log-likelihood)?
\enumerate{
\item REML  <-  FALSE is used in case of comparing models with different
“Fixed effects” (during the simplification of model)
\item REML  <-  TRUE is used in case of different random effects on the
comparing models. It is to estimate with restricted likelihood
}}
}
\value{
An object of class "lmerModLmerTest"
}
\description{
Fit a linear mixed-effects model (LMM) to data, via REML or maximum
likelihood.
}
\author{
juanprimate
}
