% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{lmerRandomList}
\alias{lmerRandomList}
\title{Analysis of variance component with maximum likelihood estimation}
\usage{
lmerRandomList(input)
}
\arguments{
\item{input}{Input a list}
}
\value{
list
\itemize{
\item covarianceMatrix
\item randFitAnova
\item contestFit
\item contest1DFit
\item contestMDFit
\item confintFitInterval
\item vcovLmer
\item VarCorrLmer
}
}
\description{
Generate analysis of variance component
with Random Factors
}
\author{
juanprimate
}
