% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semIteracao.R
\name{lmerRandomListComUma}
\alias{lmerRandomListComUma}
\title{Run Linear Mixed Model Analysis with Specific Random Effects}
\usage{
lmerRandomListComUma(input)
}
\arguments{
\item{input}{A list containing the following elements:
\describe{
\item{dataset}{A data frame with the data.}
\item{varY}{A character string representing the name of the dependent variable in the model.}
\item{REML}{A logical value indicating whether to use REML (TRUE) or maximum likelihood (FALSE) for fitting the model.}
\item{covariavel}{A character string representing the name of the covariable to include as a random effect.}
}}
}
\value{
A list containing various results from the linear mixed model analysis:
\describe{
\item{covarianceMatrixFit}{The covariance matrix of the fixed effects.}
\item{randFitAnovaFit}{The results of ANOVA on the random effects.}
\item{contestFitResp}{Results from the \code{contestFit} function.}
\item{contest1DFitResp}{Results from the \code{contest1DFit} function.}
\item{contestMDFitResp}{Results from the \code{contestMDFit} function.}
\item{confintFitIntervalResp}{Confidence intervals of the fixed effects.}
\item{vcovLmerResp}{Variance-covariance matrix of the fixed effects.}
\item{VarCorrLmerResp}{Variance components of the random effects.}
}
}
\description{
The function takes a list containing a dataset, the name of the dependent variable,
a logical value indicating whether to use REML or maximum likelihood estimation, and the
name of a covariable to include as a random effect in the model. It constructs the formula,
fits the linear mixed model, and then computes and returns various analysis results.
}
\details{
This function creates a formula for a linear mixed model with a specified random effect,
fits the model using the \code{lmer} function from the \code{lme4} package, and performs analysis
on the model using various functions. The results are returned as a list.
}
\examples{
# Example usage
# Assuming `my_data` is a data frame, "response" is the dependent variable,
# and "group" is the covariable for the random effect.
input <- list(
  dataset = my_data,
  varY = "response",
  REML = TRUE,
  covariavel = "group"
)
results <- lmerRandomListComUma(input)

}
\author{
juanprimate
}
