% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semIteracao.R
\name{lmerRandomListSemItera}
\alias{lmerRandomListSemItera}
\title{Run Mixed Model Analysis with Iterated Random Effects Formula}
\usage{
lmerRandomListSemItera(input)
}
\arguments{
\item{input}{Una lista que debe contener los siguientes elementos:
\describe{
\item{dataset}{Un data frame con los datos.}
\item{varY}{Nombre de la variable dependiente en el modelo.}
\item{REML}{Un valor lógico que indica si se debe usar REML (TRUE) o máxima verosimilitud (FALSE).}
}}
}
\value{
El resultado del análisis del modelo mixto realizado por \code{run_mixed_model_analysis}, que debe ser una lista con
varios resultados del modelo.
}
\description{
Esta función crea una fórmula para un modelo lineal mixto con efectos aleatorios para todas las variables predictores,
y luego ejecuta un análisis de modelo mixto utilizando la fórmula generada.
}
\details{
La función utiliza \code{create_formulaSemItera} para generar la fórmula del modelo lineal mixto y luego pasa esa
fórmula a \code{run_mixed_model_analysis} para ejecutar el análisis.
}
\examples{
input <- list(
  dataset = mi_dataset,
  varY = "mi_variable_dependiente",
  REML = TRUE
)
results <- lmerRandomListSemItera(input)
}
\author{
Juanprimate
}
