% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{randFitAnova}
\alias{randFitAnova}
\title{Extract and Process Random Effects from a Linear Mixed Model}
\usage{
randFitAnova(fit)
}
\arguments{
\item{fit}{An object of class "lmerModLmerTest" that represents a fitted linear mixed model.
The model should be fitted using the \code{lmer} function from the \code{lme4} package and
converted to a "lmerModLmerTest" object using \code{lmerTest::as_lmerModLmerTest}.}
}
\value{
A data frame with the processed random effects, as returned by \code{randFitAnovaParte}.
}
\description{
The function takes an object of class "lmerModLmerTest" (a fitted linear mixed model),
extracts the random effects using \code{lmerTest::rand}, and then passes these results to
\code{randFitAnovaParte} for further processing. This allows for a streamlined workflow in
analyzing and interpreting the random effects from mixed models.
}
\details{
This function extracts the random effects from a fitted linear mixed model and then processes
the results using \code{randFitAnovaParte} for further analysis.
}
\examples{
# Example of usage
# Assuming `fit` is an object of class "lmerModLmerTest"
processed_random_effects <- randFitAnova(fit)

}
\author{
juanprimate
}
