% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{randFitAnovaParte}
\alias{randFitAnovaParte}
\title{Process and Clean ANOVA Output Data}
\usage{
randFitAnovaParte(output)
}
\arguments{
\item{output}{A data frame with ANOVA results. The row names of this data frame are expected to
contain the terms of the ANOVA, which will be processed and cleaned by the function.}
}
\value{
A data frame with the same structure as the input \code{output}, but with cleaned row names.
}
\description{
The function takes the output from an ANOVA test, extracts and cleans up the row names,
and formats them into a consistent structure. The cleaned row names are then set as the row names
of the original output data frame.
}
\details{
This function processes the output from an ANOVA test, specifically focusing on
cleaning and extracting relevant information from the row names of the output data.
}
\examples{
# Example of usage
# Assuming `anova_output` is a data frame resulting from an ANOVA test
cleaned_output <- randFitAnovaParte(anova_output)

}
\author{
juanprimate
}
