% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{ranefLmer}
\alias{ranefLmer}
\title{ranef: Extract the modes of the random effects}
\usage{
ranefLmer(fit)
}
\arguments{
\item{fit}{An object of class "lmerModLmerTest".}
}
\value{
From ranef: An object of class ranef.mer composed of a list of
data frames, one for each grouping factor for the random effects. The
number of rows in the data frame is the number of levels of the grouping
factor. The number of columns is the dimension of the random effect
associated with each level of the factor.
}
\description{
A generic function to extract the conditional modes of the
random effects from a fitted model object. For linear mixed models the
conditional modes of the random effects are also the conditional means.
}
\author{
juanprimate
}
