% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{VarCorrLmer}
\alias{VarCorrLmer}
\title{Extract random effect variances from lme4 mer model object}
\usage{
VarCorrLmer(fit)
}
\arguments{
\item{fit}{An object of class "lmerModLmerTest".}
}
\value{
An object of class VarCorr.merMod or data.frame.
1.- variance
2.- sd
}
\description{
This function calculates the estimated variances, standard deviations, and
correlations between the random-effects terms in a mixed-effects model, of
class merMod (linear, generalized or nonlinear). The within-group error
variance and standard deviation are also calculated.
}
\author{
juanprimate
}
