% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmerRandom.R
\name{vcovLmer}
\alias{vcovLmer}
\title{Calculate Variance-Covariance Matrix for intercept}
\usage{
vcovLmer(fit)
}
\arguments{
\item{fit}{A fitted \link{ng}lmer model or profile}
}
\value{
A matrix of the estimated covariances between the parameter
estimates in the linear or non-linear predictor of the model.
}
\description{
Returns the variance-covariance matrix of the main parameters of intercept.
}
\author{
juanprimate
}
